DBOutputter: work around gcc 4.4.7 issue
FossilOrigin-Name: 8db5d25b8b35c9c99af009d6cf23201a994d4db0c032ab4c47daa3040c92387b

+++ qcdb/src/DBOutputter.cpp
@@ -26,8 +26,16 @@ namespace qcdb {
 			std::vector<int> const fields = comparer.get_compared_fields() ;
 
 			if(
-				std::find( fields.begin(), fields.end(), CompareFields::ePosition ) == fields.end()
-				|| std::find( fields.begin(), fields.end(), CompareFields::eAlleles ) == fields.end()
+				std::find(
+					fields.begin(),
+					fields.end(),
+					static_cast<int>(CompareFields::ePosition)
+				) == fields.end()
+				|| std::find(
+					fields.begin(),
+					fields.end(),
+					static_cast<int>(CompareFields::eAlleles)
+				) == fields.end()
 			) {
 				throw genfile::BadArgumentError(
 					"qcdb::impl::get_match_rsid()",
